requirements.txt
pandas>=2.0
PyYAML>=6.0
pytest>=7.4
